Remove getOceanCount
