+++ src/server/boards/MarsBoard.ts
@@ -1,7 +1,11 @@
+import {OCEAN_UPGRADE_TILES, TileType} from '../../common/TileType';
 import {SpaceType} from '../../common/boards/SpaceType';
 import {IPlayer} from '../IPlayer';
 import {Board} from './Board';
 import {ISpace} from './ISpace';
+import {PlacementType} from './PlacementType';
+import {AresHandler} from '../ares/AresHandler';
+import {CardName} from '../../common/cards/CardName';
 
 export class MarsBoard extends Board {
   public getCitiesOffMars(player?: IPlayer): Array<ISpace> {
@@ -22,5 +26,106 @@ export class MarsBoard extends Board {
     let greeneries = this.spaces.filter((space) => Board.isGreenerySpace(space));
     if (player !== undefined) greeneries = greeneries.filter(Board.ownedBy(player));
     return greeneries;
+  }
+
+  public getAvailableSpacesForType(player: IPlayer, type: PlacementType): ReadonlyArray<ISpace> {
+    switch (type) {
+    case 'land': return this.getAvailableSpacesOnLand(player);
+    case 'ocean': return this.getAvailableSpacesForOcean(player);
+    case 'greenery': return this.getAvailableSpacesForGreenery(player);
+    case 'city': return this.getAvailableSpacesForCity(player);
+    case 'isolated': return this.getAvailableIsolatedSpaces(player);
+    case 'volcanic': return this.getAvailableVolcanicSpaces(player);
+    case 'upgradeable-ocean': return this.getOceanSpaces({upgradedOceans: false});
+    default: throw new Error('unknown type ' + type);
+    }
+  }
+
+  /*
+   * Returns spaces on the board with ocean tiless.
+   *
+   * The default condition is to return those oceans used to count toward the global parameter, so
+   * upgraded oceans are included, but Wetlands is not. That's why the boolean values have different defaults.
+   */
+  public getOceanSpaces(include?: {upgradedOceans?: boolean, wetlands?: boolean}): ReadonlyArray<ISpace> {
+    const spaces = this.spaces.filter((space) => {
+      if (!Board.isOceanSpace(space)) return false;
+      if (space.tile?.tileType === undefined) return false;
+      const tileType = space.tile.tileType;
+      if (OCEAN_UPGRADE_TILES.has(tileType)) {
+        return include?.upgradedOceans ?? true;
+      }
+      if (tileType === TileType.WETLANDS) {
+        return include?.wetlands ?? false;
+      }
+      return true;
+    });
+    return spaces;
+  }
+
+  public getAvailableSpacesForCity(player: IPlayer): ReadonlyArray<ISpace> {
+    const spacesOnLand = this.getAvailableSpacesOnLand(player);
+    // Gordon CEO can ignore placement restrictions for Cities+Greenery
+    if (player.cardIsInEffect(CardName.GORDON)) return spacesOnLand;
+    // A city cannot be adjacent to another city
+    return spacesOnLand.filter(
+      (space) => this.getAdjacentSpaces(space).some((adjacentSpace) => Board.isCitySpace(adjacentSpace)) === false,
+    );
+  }
+
+  public getAvailableSpacesForGreenery(player: IPlayer): ReadonlyArray<ISpace> {
+    let spacesOnLand = this.getAvailableSpacesOnLand(player);
+    // Gordon CEO can ignore placement restrictions for Cities+Greenery
+    if (player.cardIsInEffect(CardName.GORDON)) return spacesOnLand;
+    // Spaces next to Red City are always unavialable.
+    if (player.game.gameOptions.pathfindersExpansion === true) {
+      spacesOnLand = spacesOnLand.filter((space) => {
+        return !this.getAdjacentSpaces(space).some((neighbor) => neighbor.tile?.tileType === TileType.RED_CITY);
+      });
+    }
+
+    const spacesForGreenery = spacesOnLand
+      .filter((space) => this.getAdjacentSpaces(space).find((adj) => adj.tile !== undefined && adj.player === player && adj.tile.tileType !== TileType.OCEAN) !== undefined);
+
+    // Spaces next to tiles you own
+    if (spacesForGreenery.length > 0) {
+      return spacesForGreenery;
+    }
+    // Place anywhere if no space owned
+    return spacesOnLand;
+  }
+
+  public getAvailableSpacesForOcean(player: IPlayer): ReadonlyArray<ISpace> {
+    return this.getSpaces(SpaceType.OCEAN, player)
+      .filter(
+        (space) => space.tile === undefined &&
+                      (space.player === undefined || space.player === player),
+      );
+  }
+
+  public getAvailableIsolatedSpaces(player: IPlayer): ReadonlyArray<ISpace> {
+    return this.getAvailableSpacesOnLand(player)
+      .filter((space: ISpace) => this.getAdjacentSpaces(space).every((space) => space.tile === undefined));
+  }
+
+  public getAvailableVolcanicSpaces(player: IPlayer): ReadonlyArray<ISpace> {
+    const volcanicSpaceIds = this.getVolcanicSpaceIds();
+
+    const spaces = this.getAvailableSpacesOnLand(player);
+    if (volcanicSpaceIds.length > 0) {
+      return spaces.filter((space) => volcanicSpaceIds.includes(space.id));
+    }
+    return spaces;
+  }
+
+  /**
+   * Almost the same as getAvailableSpacesOnLand, but doesn't apply to any player.
+   */
+  public getNonReservedLandSpaces(): ReadonlyArray<ISpace> {
+    return this.spaces.filter((space) => {
+      return (space.spaceType === SpaceType.LAND || space.spaceType === SpaceType.COVE) &&
+        (space.tile === undefined || AresHandler.hasHazardTile(space)) &&
+        space.player === undefined;
+    });
   }
 }
