+++ src/server/turmoil/parties/Reds.ts
@@ -144,7 +144,7 @@ class RedsPolicy03 implements Policy {
     if (game.marsIsTerraformed()) return false;
 
     const temperature = game.getTemperature();
-    const oceansPlaced = game.board.getOceanCount();
+    const oceansPlaced = game.board.getOceanSpaces().length;
     const oxygenLevel = game.getOxygenLevel();
     const venusScaleLevel = game.getVenusScaleLevel();
