+++ tests/cards/ares/OceanCity.spec.ts
@@ -55,7 +55,7 @@ describe('OceanCity', function() {
     expect(player.production.energy).eq(0);
     expect(player.production.megacredits).eq(3);
     expect(game.board.getCitiesOnMars()).is.empty;
-    expect(player.game.board.getCities(player).length).eq(0);
+    expect(player.game.board.getCities(player)).is.empty;
 
     action.cb(oceanSpace);
