+++ tests/cards/base/standardProjects/AquiferStandardProject.spec.ts
@@ -33,7 +33,7 @@ describe('AquiferStandardProject', function() {
   it('action', function() {
     player.megaCredits = card.cost;
     player.setTerraformRating(20);
-    expect(game.board.getOceanCount()).eq(0);
+    expect(game.board.getOceanSpaces()).is.empty;
 
     const selectSpace = cast(churnAction(card, player), SelectSpace);
     const availableSpace = selectSpace.availableSpaces[0];
@@ -44,7 +44,7 @@ describe('AquiferStandardProject', function() {
 
     expect(availableSpace.tile!.tileType).eq(TileType.OCEAN);
     expect(player.getTerraformRating()).eq(21);
-    expect(game.board.getOceanCount()).eq(1);
+    expect(game.board.getOceanSpaces()).has.length(1);
   });
 
   it('cannnot act when maximized', () => {
