+++ tests/cards/base/standardProjects/PowerPlantStandardProject.spec.ts
@@ -19,7 +19,7 @@ describe('PowerPlantStandardProjects', function() {
   it('Should act', function() {
     player.megaCredits = 11;
     player.playedCards.push(new StandardTechnology());
-    expect(game.deferredActions.length).eq(0);
+    expect(game.deferredActions).has.length;
     card.action(player);
     expect(game.deferredActions.length).eq(1);
     expect(player.megaCredits).eq(11);
