+++ tests/cards/promo/MoholeLake.spec.ts
@@ -26,7 +26,7 @@ describe('MoholeLake', function() {
     selectSpace.cb(selectSpace.availableSpaces[0]);
 
     expect(player.game.getTemperature()).to.eq(-28);
-    expect(player.game.board.getOceanCount()).eq(1);
+    expect(player.game.board.getOceanSpaces()).has.length(1);
     expect(player.getTerraformRating()).to.eq(22);
     expect(player.plants).to.eq(3);
   });
